states refactoring
