+++ examples/ag-demo-ui/src/layout/packet/sidenav/index.js
@@ -6,7 +6,8 @@ class controller {
     this.$element = $element
     this.$state = appState.$state
     this.appState = appState
-    this.sideMenuItems = [{data:{title: "Org", href: 'org'}}, {data:{title: "User", href: 'user'}}]//appState.routerStates.children
+    this.sideMenuItems = appState.routerStates.children
+      //[{data:{title: "Org", href: 'org'}}, {data:{title: "User", href: 'user'}}]//appState.routerStates.children
   }
 
   getCurrentUrl = () => this.$location.absUrl().split('#')[0].split('/')[3]
