+++ examples/ag-demo-ui/src/layout/routerStates.js
@@ -1,8 +1,12 @@
+import orgStates from '../org/states'
+import userStates from '../user/states'
+
 export const packet = {
   abstract: true,
   url:'^',
   name: 'app',
-  template: require('./packet/index.html')
+  template: require('./packet/index.html'),
+  children: [orgStates, userStates]
 }
 
 export const fresh = {
